+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-- Bundler now inlines scripts and css by default.  Pass `inlineCss` and
-  `inlineScripts` as `false` explicitly to disable.
+- BREAKING: Bundler now inlines Scripts and CSS by default.  Pass `inlineCss`
+  and `inlineScripts` as `false` explicitly to disable.
+- Fixed an issue where `exclude` option did not actually support folder names.
 
 ## 2.0.0-pre.12 - 2017-04-14
 - BREAKING: Public API change.  The `bundle()` method now takes a manifest